Split SemanticAnalyser class from UT
