+++ td/challenge260/SemanticAnalyserTest.cpp
@@ -2,64 +2,7 @@
 
 #include <stdexcept>
 
-#include "Token.hpp"
-#include "Instruction.hpp"
-
-class SemanticAnalyser
-{
-public:
-  class InvalidSemantic : public std::runtime_error
-  {
-  public:
-    InvalidSemantic(const std::string& msg) : std::runtime_error{msg}
-    {
-    }
-  };
-
-  Instructions analyse(const Tokens& tokens)
-  {
-    if (tokens.empty())
-    {
-      return {};
-    }
-    Instructions instructions;
-    unsigned alreadyProcessedTokens = 0;
-    while (tokens.size() - alreadyProcessedTokens % sizeOfInstruction && tokens.size() != alreadyProcessedTokens)
-    {
-      if (areTokensValidLdInstruction(tokens.cbegin() + alreadyProcessedTokens))
-      {
-        instructions.push_back({InstructionType::LdA, 0});
-      }
-      else if (areTokensValidOutInstruction(tokens.cbegin() + alreadyProcessedTokens))
-      {
-        instructions.push_back({InstructionType::OutA, 0});
-      }
-      else
-      {
-        throw InvalidSemantic{"Invalid instruction"};
-      }
-      alreadyProcessedTokens += sizeOfInstruction;
-    }
-    return instructions;
-  }
-
-private:
-  static constexpr auto sizeOfInstruction = 3u;
-
-  bool areTokensValidOutInstruction(Tokens::const_iterator begin)
-  {
-    return (begin++)->type == TokenType::Out
-        && (begin++)->type == TokenType::ZeroWithBrackets
-        && (begin++)->type == TokenType::A;
-  }
-
-  bool areTokensValidLdInstruction(Tokens::const_iterator begin)
-  {
-    return (begin++)->type == TokenType::Ld
-        && (begin++)->type == TokenType::A
-        && (begin++)->type == TokenType::Number8Bit;
-  }
-};
+#include "SemanticAnalyser.hpp"
 
 using namespace ::testing;
